Recycling_Segmentation/requirements.txt
python-multipart
fastapi
uvicorn
torch
transformers
safetensors
pillow
